Update MFGC_class.py
I am confused as I know you ran the code on your IDE but I cannot get the plt.show()(line 57) to work for the neural network but there are no error messages. I suspect it is some kind of looping issue as when I indent my test plot(59-63) to the same indentation as your plt.show() it stops bringing my testing graph up. However, changing the indentation of line 57 does not solve the issue.